Fixed wrong data type

+++ Sadie.Networking.Writers/Players/Inventory/PlayerInventoryRefreshWriter.cs
@@ -7,6 +7,6 @@ public class PlayerInventoryRefreshWriter : NetworkPacketWriter
 {
     public PlayerInventoryRefreshWriter()
     {
-        WriteInteger(ServerPacketId.PlayerInventoryRefresh);
+        WriteShort(ServerPacketId.PlayerInventoryRefresh);
     }
 }
